requirements.txt
python-oauth2